move ArcTo to separate file
