+++ core/src/commonTest/kotlin/dev/lennartegb/vec2compose/core/PathParserTest.kt
@@ -1,12 +1,13 @@
 package dev.lennartegb.vec2compose.core
 
+import dev.lennartegb.vec2compose.core.commands.Command
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 internal class PathParserTest {
 
     @Test
-    fun parse_path_to_list_of_commands_without_closing_end() {
+    fun `parse path to list of commands without closing end`() {
         val pathSample = "M27.05,24.55L12.15,19.65C11.85,19.35 11.483,19.2 11.05,19.2"
 
         assertEquals(
@@ -20,7 +21,7 @@ internal class PathParserTest {
     }
 
     @Test
-    fun parse_path_to_list_of_commands_with_closing_end() {
+    fun `parse path to list of commands with closing end`() {
         val pathSample = "M27.05,24.55L12.15,19.65C11.85,19.35 11.483,19.2 11.05,19.2Z"
 
         assertEquals(
